Revert "Fixed hardcoded colors"
