+++ common/src/main/java/io/novafoundation/nova/common/utils/formatting/spannable/SpannableFormatter.kt
@@ -1,6 +1,9 @@
 package io.novafoundation.nova.common.utils.formatting.spannable
 
+import android.content.Context
 import android.text.SpannedString
+import androidx.annotation.StringRes
+import io.novafoundation.nova.common.resources.ResourceManager
 import io.novafoundation.nova.common.utils.formatting.spannable.SpannableFormatter.fill
 import java.util.regex.Pattern
 
@@ -64,4 +67,14 @@ object SpannableFormatter {
 
         fun result(): CharSequence
     }
+}
+
+fun SpannableFormatter.format(resourceManager: ResourceManager, @StringRes resId: Int, vararg args: Any): SpannedString {
+    val format = resourceManager.getString(resId)
+    return format(format, *args)
+}
+
+fun SpannableFormatter.format(context: Context, @StringRes resId: Int, vararg args: Any): SpannedString {
+    val format = context.getString(resId)
+    return format(format, *args)
 }
