+++ feature-account-impl/src/main/java/io/novafoundation/nova/feature_account_impl/presentation/paritySigner/config/PolkadotVaultVariantConfigBuilder.kt
@@ -32,8 +32,6 @@ interface PolkadotVaultVariantConfigBuilder {
 
             fun step(@StringRes contentRes: Int)
 
-            fun step(content: CharSequence)
-
             fun image(@StringRes labelRes: Int, @DrawableRes imageRes: Int)
         }
     }
@@ -120,19 +118,19 @@ private class RealInstructionsBuilder(
 
     override fun step(contentRes: Int) {
         val content = resourceManager.getText(contentRes)
-        step(content)
-    }
-
-    override fun step(content: CharSequence) {
-        stepsCounter += 1
-
-        val stepInstruction = Instruction.Step(stepsCounter, content)
-        instructions.add(stepInstruction)
+        addStep(content)
     }
 
     override fun image(labelRes: Int, imageRes: Int) {
         val imageInstruction = Instruction.Image(resourceManager.getString(labelRes), imageRes)
         instructions.add(imageInstruction)
+    }
+
+    private fun addStep(content: CharSequence) {
+        stepsCounter += 1
+
+        val stepInstruction = Instruction.Step(stepsCounter, content)
+        instructions.add(stepInstruction)
     }
 
     fun build(): List<Instruction> {
