+++ feature-ledger-impl/src/main/res/layout/fragment_generic_import_ledger_start.xml
@@ -32,14 +32,14 @@
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginTop="16dp"
-                android:text="@string/account_ledger_import_start_title"
-                android:textColor="@color/text_primary" />
+                android:textColor="@color/text_primary"
+                android:text="@string/account_ledger_import_start_title" />
 
             <io.novafoundation.nova.common.view.LinkView
-                android:id="@+id/startImportGenericLedgerGuideLink"
                 android:layout_width="wrap_content"
-                android:layout_height="match_parent"
+                android:id="@+id/startImportGenericLedgerGuideLink"
                 android:layout_marginTop="8dp"
+                android:layout_height="match_parent"
                 app:linkText="@string/account_ledger_import_start_guide" />
 
             <io.novafoundation.nova.common.view.NovaConnectView
@@ -62,7 +62,8 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:layout_marginTop="24dp"
-                app:stepNumber="2" />
+                app:stepNumber="2"
+                app:stepText="@string/account_ledger_generic_import_start_step_2" />
 
             <io.novafoundation.nova.common.view.InstructionStepView
                 android:id="@+id/startImportGenericLedgerStep3"
@@ -70,7 +71,8 @@
                 android:layout_height="wrap_content"
                 android:layout_marginTop="24dp"
                 app:layout_constraintEnd_toEndOf="parent"
-                app:stepNumber="3" />
+                app:stepNumber="3"
+                app:stepText="@string/account_ledger_import_start_step_3" />
 
             <io.novafoundation.nova.common.view.InstructionStepView
                 android:id="@+id/startImportGenericLedgerStep4"
@@ -78,7 +80,8 @@
                 android:layout_height="wrap_content"
                 android:layout_marginTop="24dp"
                 app:layout_constraintEnd_toEndOf="parent"
-                app:stepNumber="4" />
+                app:stepNumber="4"
+                app:stepText="@string/account_ledger_import_start_step_4" />
         </LinearLayout>
 
     </ScrollView>
