+++ feature-ledger-impl/src/main/res/layout/fragment_import_ledger_start.xml
@@ -32,14 +32,14 @@
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginTop="16dp"
-                android:text="@string/account_ledger_import_start_title"
-                android:textColor="@color/text_primary" />
+                android:textColor="@color/text_primary"
+                android:text="@string/account_ledger_import_start_title" />
 
             <io.novafoundation.nova.common.view.LinkView
-                android:id="@+id/startImportLedgerGuideLink"
                 android:layout_width="wrap_content"
-                android:layout_height="match_parent"
+                android:id="@+id/startImportLedgerGuideLink"
                 android:layout_marginTop="8dp"
+                android:layout_height="match_parent"
                 app:linkText="@string/account_ledger_import_start_guide" />
 
             <io.novafoundation.nova.common.view.NovaConnectView
@@ -54,14 +54,16 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:layout_marginTop="32dp"
-                app:stepNumber="1" />
+                app:stepNumber="1"
+                app:stepText="@string/account_ledger_import_start_step_1" />
 
             <io.novafoundation.nova.common.view.InstructionStepView
                 android:id="@+id/startImportLedgerStep2"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:layout_marginTop="24dp"
-                app:stepNumber="2" />
+                app:stepNumber="2"
+                app:stepText="@string/account_ledger_import_start_step_2" />
 
             <io.novafoundation.nova.common.view.InstructionStepView
                 android:id="@+id/startImportLedgerStep3"
@@ -69,7 +71,8 @@
                 android:layout_height="wrap_content"
                 android:layout_marginTop="24dp"
                 app:layout_constraintEnd_toEndOf="parent"
-                app:stepNumber="3" />
+                app:stepNumber="3"
+                app:stepText="@string/account_ledger_import_start_step_3" />
 
             <io.novafoundation.nova.common.view.InstructionStepView
                 android:id="@+id/startImportLedgerStep4"
@@ -77,19 +80,20 @@
                 android:layout_height="wrap_content"
                 android:layout_marginTop="24dp"
                 app:layout_constraintEnd_toEndOf="parent"
-                app:stepNumber="4" />
+                app:stepNumber="4"
+                app:stepText="@string/account_ledger_import_start_step_4" />
 
             <io.novafoundation.nova.common.view.AlertView
-                android:id="@+id/startImportLedgerDepractionWarning"
                 android:layout_width="match_parent"
-                android:layout_height="wrap_content"
                 android:layout_marginTop="24dp"
+                android:layout_height="wrap_content"
                 android:text="@string/account_ledger_legacy_warning_title"
-                android:visibility="gone"
-                app:AlertView_action="@string/common_find_out_more"
-                app:AlertView_description="@string/account_ledger_legacy_warning_message"
                 app:alertMode="warning"
-                tools:visibility="visible" />
+                android:visibility="gone"
+                tools:visibility="visible"
+                android:id="@+id/startImportLedgerDepractionWarning"
+                app:AlertView_description="@string/account_ledger_legacy_warning_message"
+                app:AlertView_action="@string/common_find_out_more" />
         </LinearLayout>
 
     </ScrollView>
